backend/requirements.txt
Django>=3.2.9,<3.3.0
#최신 버전?
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.9.2,<3.0.0
djangorestframework-simplejwt==5.0.0